# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bookd API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
  message: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  CUSTOMER: "CUSTOMER",
  PROVIDER: "PROVIDER",
} as const;

export interface User {
  id: string;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  isActive: boolean;
  createdAt: string;
}

export type RegisterInputRole =
  (typeof RegisterInputRole)[keyof typeof RegisterInputRole];

export const RegisterInputRole = {
  CUSTOMER: "CUSTOMER",
  PROVIDER: "PROVIDER",
} as const;

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
  role: RegisterInputRole;
  phone?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface RefreshInput {
  refreshToken: string;
}

export interface AuthResponse {
  token: string;
  refreshToken: string;
  user: User;
}

export type ServiceCategory =
  (typeof ServiceCategory)[keyof typeof ServiceCategory];

export const ServiceCategory = {
  HEALTH: "HEALTH",
  BEAUTY: "BEAUTY",
  FITNESS: "FITNESS",
  EDUCATION: "EDUCATION",
  HOME: "HOME",
  OTHER: "OTHER",
} as const;

export interface Service {
  id: string;
  providerId: string;
  provider: User;
  title: string;
  description: string;
  category: ServiceCategory;
  price: number;
  durationMinutes: number;
  location: string;
  /** @nullable */
  imageUrl?: string | null;
  isActive: boolean;
  /** @nullable */
  averageRating?: number | null;
  reviewCount: number;
  createdAt: string;
}

export interface ServiceListResponse {
  services: Service[];
  total: number;
  page: number;
  size: number;
  totalPages: number;
}

export type ServiceInputCategory =
  (typeof ServiceInputCategory)[keyof typeof ServiceInputCategory];

export const ServiceInputCategory = {
  HEALTH: "HEALTH",
  BEAUTY: "BEAUTY",
  FITNESS: "FITNESS",
  EDUCATION: "EDUCATION",
  HOME: "HOME",
  OTHER: "OTHER",
} as const;

export interface ServiceInput {
  title: string;
  description: string;
  category: ServiceInputCategory;
  price: number;
  durationMinutes: number;
  location: string;
  imageUrl?: string;
}

export type ServiceUpdateCategory =
  (typeof ServiceUpdateCategory)[keyof typeof ServiceUpdateCategory];

export const ServiceUpdateCategory = {
  HEALTH: "HEALTH",
  BEAUTY: "BEAUTY",
  FITNESS: "FITNESS",
  EDUCATION: "EDUCATION",
  HOME: "HOME",
  OTHER: "OTHER",
} as const;

export interface ServiceUpdate {
  title?: string;
  description?: string;
  category?: ServiceUpdateCategory;
  price?: number;
  durationMinutes?: number;
  location?: string;
  imageUrl?: string;
  isActive?: boolean;
}

export interface Availability {
  id: string;
  serviceListingId: string;
  slotDate: string;
  startTime: string;
  endTime: string;
  isBooked: boolean;
}

export type AvailabilityInputSlotsItem = {
  startTime: string;
  endTime: string;
};

export interface AvailabilityInput {
  serviceListingId: string;
  startDate: string;
  endDate: string;
  slots: AvailabilityInputSlotsItem[];
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  PENDING: "PENDING",
  CONFIRMED: "CONFIRMED",
  CANCELLED: "CANCELLED",
  COMPLETED: "COMPLETED",
} as const;

export interface Booking {
  id: string;
  customerId: string;
  customer?: User;
  serviceListingId?: string;
  service: Service;
  availabilityId: string;
  availability: Availability;
  status: BookingStatus;
  totalPrice: number;
  /** @nullable */
  notes?: string | null;
  hasReview: boolean;
  createdAt: string;
}

export interface BookingInput {
  serviceListingId: string;
  availabilityId: string;
  notes?: string;
}

export interface BookingListResponse {
  bookings: Booking[];
  total: number;
  page: number;
  size: number;
  totalPages: number;
}

export interface Review {
  id: string;
  bookingId: string;
  customerId: string;
  customer: User;
  serviceListingId: string;
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}

export interface ReviewInput {
  bookingId: string;
  rating: number;
  comment?: string;
}

export interface ReviewListResponse {
  reviews: Review[];
  total: number;
  page: number;
  totalPages: number;
}

export interface WeeklyRevenue {
  week: string;
  revenue: number;
}

export interface ProviderStats {
  totalBookings: number;
  totalRevenue: number;
  /** @nullable */
  averageRating?: number | null;
  bookingsThisMonth: number;
  activeServices: number;
  revenueByWeek: WeeklyRevenue[];
}

export interface Notification {
  id: string;
  userId: string;
  type: string;
  message: string;
  isRead: boolean;
  createdAt: string;
}

export interface UserUpdate {
  name?: string;
  phone?: string;
  avatarUrl?: string;
}

export interface PasswordChangeInput {
  currentPassword: string;
  newPassword: string;
}

export type ListServicesParams = {
  category?: string;
  minPrice?: number;
  maxPrice?: number;
  location?: string;
  search?: string;
  sort?: string;
  page?: number;
  size?: number;
};

export type GetServiceAvailabilityParams = {
  serviceId: string;
  date?: string;
};

export type ListMyBookingsParams = {
  status?: string;
  page?: number;
  size?: number;
};

export type ListServiceReviewsParams = {
  serviceId: string;
  page?: number;
  size?: number;
};

export type ListProviderBookingsParams = {
  status?: string;
  page?: number;
  size?: number;
};