# lib/api-zod/src/generated/types/passwordChangeInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bookd API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PasswordChangeInput {
  currentPassword: string;
  newPassword: string;
}